Lib 180 fetching main books and my books page (#112)
* add infinite scroll to main catalogue and my books page

* fixed
