+++ src/components/books-list/books-list-results.js
@@ -21,7 +21,6 @@ import { trimmedString } from './../../utils/functions/trim-long-string';
 import { types } from '../../types';
 
 const BooksListResults = ({ books, isStartedSearch }) => {
-    const sortedBooks = books.sort((a, b) => (a.id < b.id ? 1 : -1));
     return (
         <Card>
             <PerfectScrollbar>
@@ -42,7 +41,7 @@ const BooksListResults = ({ books, isStartedSearch }) => {
                         </TableHead>
                         <TableBody>
                             {books.length ? (
-                                sortedBooks.map((book) => {
+                                books.map((book) => {
                                     return (
                                         <TableRow
                                             onClick={() =>
