+++ src/pages/main-catalogue.js
@@ -1,37 +1,13 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import { Typography } from '@mui/material';
 import { DashboardLayout } from '../components/dashboard-layout';
 import { BooksAPI } from '../api/books-api';
 import BooksCatalogue from '../components/books-catalogue';
-import { useCustomSnackbar } from '../utils/custom-snackbar-hook';
-import { useRouter } from 'next/router';
-import { LOGIN_PATH } from '../common/constants/route-constants';
+import { useInfiniteScroll } from '../utils/infinite-scroll-hook';
 
 const MainCatalogue = () => {
-    const router = useRouter();
     const [books, setBooks] = useState([]);
-    const [isLoaded, setIsLoaded] = useState(false);
-    const { defaultErrorSnackbar } = useCustomSnackbar();
-
-    const updateBooks = (booksList) => {
-        setBooks(booksList);
-    };
-
-    useEffect(() => {
-        BooksAPI.getAllBooks()
-            .then((res) => {
-                setBooks(res.data);
-                setIsLoaded(true);
-            })
-            .catch((err) => {
-                if (err.response.status === 403) {
-                    router.replace(LOGIN_PATH);
-                    localStorage.removeItem('token');
-                } else {
-                    defaultErrorSnackbar();
-                }
-            });
-    }, [defaultErrorSnackbar, router]);
+    const { isLoaded, setIsLoaded } = useInfiniteScroll(BooksAPI.getAllBooks, books, setBooks, 15);
 
     if (!isLoaded) {
         return (
@@ -44,7 +20,7 @@ const MainCatalogue = () => {
             <BooksCatalogue
                 books={books}
                 title={'Main catalogue'}
-                onUpdateBooks={updateBooks}
+                onUpdateBooks={setBooks}
                 onUpdateLoadingStatus={setIsLoaded}
                 isAllBooks={true}
             />
