+++ src/pages/my-books.js
@@ -1,41 +1,18 @@
-import { useState, useEffect } from 'react';
-import { useRouter } from 'next/router';
+import { useState } from 'react';
 import { Typography } from '@mui/material';
 import { DashboardLayout } from '../components/dashboard-layout';
 import BooksCatalogue from '../components/books-catalogue';
 import { BooksAPI } from '../api/books-api';
-import { useCustomSnackbar } from '../utils/custom-snackbar-hook';
-import { LOGIN_PATH } from '../common/constants/route-constants';
+import { useInfiniteScroll } from './../utils/infinite-scroll-hook';
 
 const OwnerCatalogue = () => {
-    const router = useRouter();
     const [books, setBooks] = useState([]);
-    const [isLoaded, setIsLoaded] = useState(false);
-    const { enqueueSnackbar, defaultErrorSnackbar } = useCustomSnackbar();
-
-    const updateBooks = (booksList) => {
-        setBooks(booksList);
-    };
-
-    const updateLoadingStatus = () => {
-        setIsLoaded(true);
-    };
-
-    useEffect(() => {
-        BooksAPI.getOwnerBooks()
-            .then((res) => {
-                setBooks(res.data);
-                setIsLoaded(true);
-            })
-            .catch((err) => {
-                if (err.response.status === 403) {
-                    router.replace(LOGIN_PATH);
-                    localStorage.removeItem('token');
-                } else {
-                    defaultErrorSnackbar();
-                }
-            });
-    }, [defaultErrorSnackbar, enqueueSnackbar, router]);
+    const { isLoaded, setIsLoaded } = useInfiniteScroll(
+        BooksAPI.getOwnerBooks,
+        books,
+        setBooks,
+        15
+    );
 
     if (!isLoaded) {
         return (
@@ -48,8 +25,8 @@ const OwnerCatalogue = () => {
             <BooksCatalogue
                 books={books}
                 title={'My books'}
-                onUpdateBooks={updateBooks}
-                onUpdateLoadingStatus={updateLoadingStatus}
+                onUpdateBooks={setBooks}
+                onUpdateLoadingStatus={setIsLoaded}
                 isMyBooks={true}
             />
         );
